ci(linter-configurations): configure linter for the project (#18)
The line length of characters rule was set to 80 characters but now I changed it to 100 characters.  All our java code followed the 100 characters per line rule. So, the linter must check the code with 100 characters per line. Allow the empty methods, constructors, and catches.  Configure the check to enforce final parameters only for constructors. Allow ignoring primitive datatypes as parameters. configure the check so that it checks local variables but not parameters.

closes #17